Feat: updatePlanStatus 함수 추가
- useChangePlanStatus hook에서 handleChangePlanStatus 함수 제거,
  컴포넌트에 선언
