+++ src/components/common/button/ChangeStatusButton.tsx
@@ -9,7 +9,7 @@ interface Props extends ButtonCompProps {
 }
 
 const ChangeStatusButton = (props: Props) => {
-  const { value, type, ariaLable, name, leftText, disabled, onClick } = props;
+  const { value, type, name, leftText, disabled, onClick } = props;
 
   return (
     <div
@@ -26,6 +26,7 @@ const ChangeStatusButton = (props: Props) => {
       </p>
       <button
         type={type}
+        name={name}
         disabled={disabled}
         onClick={onClick}
         className="flex-box p-3 border rounded-lg font-semibold border-blue text-blue hover:bg-blue_light_1 duration-200 disabled:border-gray_dark_1 disabled:cursor-default disabled:bg-gray_light_3 disabled:text-gray_dark_1
