+++ src/components/plan/changePlanStatus/ChangePlanStatus.tsx
@@ -2,8 +2,11 @@
 
 import React from 'react';
 
+import { useRouter } from 'next/navigation';
+
 import ChangeStatusButton from '@/components/common/button/ChangeStatusButton';
 import useChangePlanStatus from '@/hooks/useChangePlanStatus';
+import useConfirm from '@/hooks/useConfirm';
 
 import type { PlanStatus } from '@/types/aboutPlan.type';
 
@@ -13,10 +16,36 @@ interface Props {
 }
 
 const ChangePlanStatus = ({ status, planId }: Props) => {
-  const { leftText, disabled, handleChangePlanStatus } = useChangePlanStatus({
+  const { leftText, disabled } = useChangePlanStatus({
     status,
     planId,
   });
+
+  const router = useRouter();
+  const confirm = useConfirm();
+
+  const handleChangePlanStatus = () => {
+    if (status === 'planning') {
+      const confTitle = '여행 중으로 변경';
+      const confDesc =
+        '여행 중으로 변경할 경우 다시 계획 중으로 되돌릴 수 없습니다. 변경하시겠습니까?';
+
+      const confFunc = async () => {
+        // changeStateMutate([planId, 'traveling']);
+        // scrollTop();
+      };
+      confirm.default(confTitle, confDesc, confFunc);
+    } else {
+      const confTitle = '여행 완료로 변경';
+      const confDesc =
+        '여행을 완료하시면 더 이상 여행 내용을 수정하실 수 없습니다. 완료하시겠습니까?';
+      const confFunc = () => {
+        // changeStateMutate([planId, 'recording']);
+        router.push(`/addPhoto/${planId}`);
+      };
+      confirm.default(confTitle, confDesc, confFunc);
+    }
+  };
 
   return (
     <div
